Implement usb port redirection abusing UsbDk_StartRedirect API.
When device inserted to usb port IOCTL_USBDK_ADD_REDIRECT will return
HANDLE to this device, so userspace can communicate with the captured
device through this handle. When device is not inserted IOCTL_USBDK_ADD_REDIRECT
returns handle to stub (port control device), it means that userspace should wait
when device will be inserted to redirected port. Once device inserted it will be
captured by viusbdk driver. In this case user application need to find symlink to
device filter associated with captured device and open it, the returned handle
should be setted with UsbDk_SetRedirectorSystemHandle.